requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
gspread==5.12.0
google-auth-oauthlib==1.1.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
anthropic==0.28.0
openai==1.3.9
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
pytz==2023.3
